moved ResponseMocker to proxy types
