+++ src/act.type.ts
@@ -1,4 +1,4 @@
-import { Mockapi, Moctokit } from "@kie/mock-github"
+import { ResponseMocker } from "@aj/proxy/proxy.types";
 
 export type Workflow = {
   jobId: string;
@@ -22,11 +22,3 @@ export type RunOpts = {
   };
   mockApi?: ResponseMocker[];
 };
-
-export type ResponseMocker = ReturnType<typeof Mockapi.prototype.mock["any"]["any"]["any"]> | ReturnType<Extract<typeof Moctokit.prototype.rest>>;
-
-type Extract<T extends typeof Moctokit.prototype.rest> = {
-  [K in keyof T]:  {
-    [W in keyof T[K]]: T[K][W]
-  }[keyof T[K]]
-}[keyof T];
